refactor: InputContainer.tsx 파일 컴포넌트로 이동
