+++ src/components/form/index.ts
@@ -1,2 +1,3 @@
 export { default as Button } from './Button';
 export { default as Input } from './Input';
+export { default as InputContainer } from './InputContainer';
