+++ src/container/RegisterCard/CardHolderContainer.tsx
@@ -1,6 +1,5 @@
-import { InputContainer } from '../index';
+import { Input, InputContainer } from '../../components/form';
 import { useInput } from '../../hooks';
-import { Input } from '../../components/form';
 import { RegisterCardType } from '../../pages/RegisterCard';
 import { useEffect } from 'react';
