--- src/container/index.ts
@@ -1,1 +1,0 @@
-export { default as InputContainer } from './InputContainer';
